Add tratamento de exceção banco de dados

+++ src/main/java/com/desuperior/dscommerce/services/ProductService.java
@@ -3,11 +3,14 @@ package com.desuperior.dscommerce.services;
 import com.desuperior.dscommerce.dto.ProductDTO;
 import com.desuperior.dscommerce.entities.Product;
 import com.desuperior.dscommerce.repositories.ProductRepository;
+import com.desuperior.dscommerce.services.exceptions.DatabaseException;
 import com.desuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -52,12 +55,20 @@ public class ProductService {
         return new ProductDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void  delete(Long id) {
-        repository.deleteById(id);
+
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Product not found with id " + id);
+        }
+        try{
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Integrity violation");
+        }
+
+
     }
-
-
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
